Corrige Lista 04
Questões 01, 02, 03, 07, 08, 17, 19, 20

+++ listas/04-lacos/questao-17.rb
@@ -1,11 +1,11 @@
 tempo = gets.to_i
-inicio = gets.to_i
+inicio = gets.to_f
 massa_final = inicio.to_f
 duracao = 0
 
 while massa_final >= 0.5
   duracao += tempo
-  massa_final /= 2
+  massa_final /= 2.0
 end
 
 h = duracao / 3600
